removed blitz library template project
