--- BlitzLibrary/build.properties
@@ -1,11 +1,0 @@
-# Build information
-out=FRCUserProgram
-src.dir=src
-build.dir=build
-out.exe=Debug/${out}
-
-# Simulation
-simulation.world.file=/usr/share/frcsim/worlds/GearsBotDemo.world
-
-# Use the current C++ library by default
-cpp-version=current
