--- BlitzLibrary/src/DriveTrains/Arcade/ArcadeDrive.hpp
@@ -1,15 +1,0 @@
-/*
- * ArcadeDrive.hpp
- *
- *  Created on: Aug 6, 2018
- *      Author: Sam
- */
-
-#ifndef SRC_DRIVETRAINS_ARCADE_ARCADEDRIVE_HPP_
-#define SRC_DRIVETRAINS_ARCADE_ARCADEDRIVE_HPP_
-
-
-
-
-
-#endif /* SRC_DRIVETRAINS_ARCADE_ARCADEDRIVE_HPP_ */
