--- BlitzLibrary/src/DriveTrains/DriveInterface.cpp
@@ -1,10 +1,0 @@
-/*
- * DriveInterface.cpp
- *
- *  Created on: Aug 6, 2018
- *      Author: Sam
- */
-
-
-
-
