--- BlitzLibrary/src/DriveTrains/DriveInterface.hpp
@@ -1,15 +1,0 @@
-/*
- * DriveInterface.hpp
- *
- *  Created on: Aug 6, 2018
- *      Author: Sam
- */
-
-#ifndef SRC_DRIVETRAINS_DRIVEINTERFACE_HPP_
-#define SRC_DRIVETRAINS_DRIVEINTERFACE_HPP_
-
-
-
-
-
-#endif /* SRC_DRIVETRAINS_DRIVEINTERFACE_HPP_ */
