--- BlitzLibrary/src/DriveTrains/Mecanum/MecanumDrive.hpp
@@ -1,15 +1,0 @@
-/*
- * MecanumDrive.hpp
- *
- *  Created on: Aug 6, 2018
- *      Author: Sam
- */
-
-#ifndef SRC_DRIVETRAINS_MECANUM_MECANUMDRIVE_HPP_
-#define SRC_DRIVETRAINS_MECANUM_MECANUMDRIVE_HPP_
-
-
-
-
-
-#endif /* SRC_DRIVETRAINS_MECANUM_MECANUMDRIVE_HPP_ */
