--- BlitzLibrary/src/DriveTrains/Octocanum/OctocanumDrive.hpp
@@ -1,15 +1,0 @@
-/*
- * OctocanumDrive.hpp
- *
- *  Created on: Aug 6, 2018
- *      Author: Sam
- */
-
-#ifndef SRC_DRIVETRAINS_OCTOCANUM_OCTOCANUMDRIVE_HPP_
-#define SRC_DRIVETRAINS_OCTOCANUM_OCTOCANUMDRIVE_HPP_
-
-
-
-
-
-#endif /* SRC_DRIVETRAINS_OCTOCANUM_OCTOCANUMDRIVE_HPP_ */
