--- BlitzLibrary/src/DriveTrains/Tank/Tank.hpp
@@ -1,15 +1,0 @@
-/*
- * Tank.hpp
- *
- *  Created on: Aug 6, 2018
- *      Author: Sam
- */
-
-#ifndef SRC_DRIVETRAINS_TANK_TANK_HPP_
-#define SRC_DRIVETRAINS_TANK_TANK_HPP_
-
-
-
-
-
-#endif /* SRC_DRIVETRAINS_TANK_TANK_HPP_ */
