--- BlitzLibrary/src/Robot.cpp
@@ -1,27 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2017-2018 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-#include <WPILib.h>
-#include <Joystick.h>
-#include <Spark.h>
-
-/**
- * This is a demo program showing the use of the DifferentialDrive class.
- * Runs the motors with arcade steering.
- */
-class Robot : public SampleRobot
-{
-
-
-public:
-	void TeleopPeriodic()
-	{
-
-	}
-};
-
-START_ROBOT_CLASS(Robot)
