+++ BlitzLibrary/src/main/cpp/BlitzLogger.cpp
