+++ BlitzLibrary/src/main/include/BlitzLogger.hpp
