--- BlitzLibraryWorkspace.code-workspace
@@ -1,8 +1,0 @@
-{
-	"folders": [
-		{
-			"path": "Blitz-Library-Template"
-		}
-	],
-	"settings": {}
-}
